Bug 370: Can use wrong password to login BMC
The value of the -g option of useradd(8) follows the wrong format of
shadow(5)/crypt(3). It uses only the first 8 characters of inputed
password and hence fails to block wrong passwords whose first 8
characters happen to be the same with the correct password.

By using the glibc2 version of this function, we make it support
arbitrary length of passwords.

Change-Id: If0687e311fa056839fcccd31734e4dcf0b7a7fac
Signed-off-by: Wei-Chung Wen <wei-chung.wen@foxconn.com>