obmc-ast-watchdog: rewrite to support both heartbeat and PID files
In the case of heartbeat file, this program will periodically recreate
heartbeat file, and service is expected to delete this file to prove its
responsiveness.

In the case of PID file, service is considered responsive as long as
/proc/PID exists.

A service will be considered dead if it is non-responsive more than 10
successive checks.

This program checks every 5 seconds whether all services (it cares) are
alive. Only when all services are alive will WDT1 be kicked.

Currently, 3 services are checked:
1. Redfish
2. Hwmon
3. Event daemon

Change-Id: Iae68caf8b7090cb1aeb92bca5731ddf00889f611
Signed-off-by: Wei-Chung Wen <wei-chung.wen@foxconn.com>