Use OpenPower variable for PS derating factor
The openpower-occ-control code needs to know the power
supply derating factor, but will default to a value if
one isn't supplied externally in do_configure.

Since other will also need to use that same value, use
a common variable for it defined in openpower.inc, and
now pass it into the configure step.

This derating factor is an OpenPower concept used by the
Power processor's OCC thermal control subsystem.

Tested: Check that the derating factor variable in config.h
        in the openenpower-occ-control repo follows the
        variable in openpower.inc.

Change-Id: I259b6086ebe70b2ac6eccdd244e43a7d36a28a77
Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Signed-off-by: Doyle Huang <doyle.sy.huang@mail.foxconn.com>